rework test to use TCC
This. Friggin. Thing.  was issuing a warning from YAML::XS on some
recent version of perl and or YAML::XS.  Anyway, on closer inspection
realized that it was generating a config file with YAML::XS, but this
is better handled by TCC now.  On rewriting the test for TCC, it
was issueing other warnings from Rose::Planter.  Because. That. Is.
What. It. Does.  *sigh*.  ignore the error via capture_stderr.

Better. Little.  Yellow.  Awesome.