version dirctory for app template
from now on we will add a version to the path to the generate templates
because if we need to delete files there is no easy way to do that for
already installed dists.

Do not need to create a new version directory for minor changes to the
template, just when we add or remove a file.